+++ Sigil/Emit.StoreField.cs
@@ -38,7 +38,7 @@ namespace Sigil
 
                 if (unaligned.HasValue)
                 {
-                    UpdateState(OpCodes.Unaligned, unaligned.Value, StackTransition.None().Wrap("StoreField"));
+                    UpdateState(OpCodes.Unaligned, (byte)unaligned.Value, StackTransition.None().Wrap("StoreField"));
                 }
 
                 var transitions =
