+++ Sigil/Emit.StoreIndirect.cs
@@ -36,7 +36,7 @@ namespace Sigil
 
             if (unaligned.HasValue)
             {
-                UpdateState(OpCodes.Unaligned, unaligned.Value, StackTransition.None().Wrap("StoreIndirect"));
+                UpdateState(OpCodes.Unaligned, (byte)unaligned.Value, StackTransition.None().Wrap("StoreIndirect"));
             }
 
             if (type.IsPointer)
